Work in progress: Relationship to User

+++ Python/Energy/nest-energy-api/src/supply-points/supply-points.service.ts
@@ -1,32 +1,60 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository, DeleteResult, Connection } from 'typeorm';
-import { SupplyPoint } from '../entities/supply-point.entity';
+import { SupplyPointEntity } from '../entities/supply-point.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { UserEntity } from 'src/users/user-entity';
+import { SupplyPointRO } from 'src/entities/supply-point.ro';
 
 @Injectable()
 export class SupplyPointsService {
 
   constructor(
-    @InjectRepository(SupplyPoint) private readonly supplyPointRepository: Repository<SupplyPoint>) {
+    @InjectRepository(SupplyPointEntity) private readonly supplyPointRepository: Repository<SupplyPointEntity>,
+    @InjectRepository(UserEntity) private readonly userRepository: Repository<UserEntity>) {
   }
 
-  async findAll(): Promise<SupplyPoint[]> {
-    return await this.supplyPointRepository.find();
+  async findAll(userId: number): Promise<SupplyPointRO[]> {
+    const sps = await this.supplyPointRepository.find({ where: { userId } });
+    return sps.map(sp => this.toResponseObject(sp));
   }
 
-  async find(id: number): Promise<SupplyPoint> {
-    return await this.supplyPointRepository.findOne(id, { relations: ['measuredValues'] });
+  async find(userId: number, id: number): Promise<SupplyPointRO> {
+    const sp = await this.supplyPointRepository.findOne(id, { relations: ['measuredValues'] });
+    return this.toResponseObject(sp);
   }
 
-  async create(supplyPoint: SupplyPoint): Promise<SupplyPoint> {
-    return this.supplyPointRepository.save(supplyPoint);
+  async create(userId: number, supplyPoint: SupplyPointEntity): Promise<SupplyPointRO> {
+    const user = await this.userRepository.findOne(userId);
+    const sp = await this.supplyPointRepository.save({ ...supplyPoint, user });
+    return this.toResponseObject(sp);
   }
 
-  async update(supplyPoint: SupplyPoint): Promise<SupplyPoint> {
-    return await this.supplyPointRepository.save(supplyPoint);
+  async update(userId: number, supplyPoint: SupplyPointEntity): Promise<SupplyPointRO> {
+    let sp = await this.supplyPointRepository.findOne(supplyPoint.id);
+    if (!sp) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+    this.ensureOwnership(sp, userId);
+    sp = await this.supplyPointRepository.save(supplyPoint);
+    return this.toResponseObject(sp);
   }
 
-  async delete(id: number): Promise<DeleteResult> {
+  async delete(userId: number, id: number): Promise<DeleteResult> {
+    const sp = await this.supplyPointRepository.findOne(id);
+    if (!sp) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+    this.ensureOwnership(sp, userId);
     return await this.supplyPointRepository.delete(id);
+  }
+
+  private toResponseObject(supplyPoint: SupplyPointEntity): SupplyPointRO {
+    return { ...supplyPoint, user: supplyPoint.user.toResponseObject() };
+  }
+
+  private ensureOwnership(supplyPoint: SupplyPointEntity, userId: number): void {
+    if (supplyPoint.userId !== userId) {
+      throw new HttpException('Invalid user', HttpStatus.UNAUTHORIZED);
+    }
   }
 }
